+++ CHANGELOG.md
@@ -11,7 +11,8 @@ All notable changes to `Wildfire` will be documented in this file.
 
 ## [0.5.0](https://github.com/rougin/wildfire/compare/v0.4.0...v0.5.0) - 2018-09-23
 
-**NOTE**: This release will introduce a backward compatability break if upgrading from `v0.4.0` release.
+> [!NOTE]
+> This release will introduce a backward compatability break if upgrading from `v0.4.0` release.
 
 ### Added
 - `Model` class