+++ CHANGELOG.md
@@ -5,13 +5,13 @@ All notable changes to `Wildfire` will be documented in this file.
 ## [0.6.0](https://github.com/rougin/wildfire/compare/v0.5.0...v0.6.0) - Unreleased
 
 ### Added
-- `$pagee` variable for pagination customization in `PaginateTrait`
-- `init` in `WildfireTrait` to initialize `Wildfire` without specifying
+- `$pagee` for customizing pagination in `PaginateTrait`
 
 ### Changed
+- `WildfireTrait::wildfire` returns the `Wildfire` instance
+- Code coverage from `Scrutinizer CI` to `Codecov`
 - Code documentation by `php-cs-fixer`, code quality by `phpstan`
 - Workflow from `Travis CI` to `Github Actions`
-- Code coverage from `Scrutinizer CI` to `Codecov`
 
 ### Fixed
 - Getting offset value in `PaginateTrait`